api/requirements.txt
# Core API
fastapi==0.112.0
uvicorn[standard]==0.30.6
httpx==0.27.0
requests==2.31.0
pyyaml==6.0.1

# Data / ML
pandas==2.2.2
numpy==1.26.4
scikit-learn
joblib

# AI (optional — works without API key)
openai>=1.0.0

# ---- Phase 2A additions ----

# Database (PostgreSQL + async)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.
bcrypt==4.0.1
pydantic[email]>=2.0.0

# Caching
redis[hiredis]>=5.0.0
